Some changes for performance. Signed-off-by:Phillip Kruger <phillip.kruger@gmail.com>
