+++ server/implementation/src/main/java/io/smallrye/graphql/spi/ClassloadingService.java
@@ -19,11 +19,12 @@ import io.smallrye.graphql.execution.Classes;
  */
 public interface ClassloadingService {
 
+    ServiceLoader<ClassloadingService> classloadingServices = ServiceLoader.load(ClassloadingService.class);
+
     static ClassloadingService load() {
         ClassloadingService classloadingService;
         try {
-            ServiceLoader<ClassloadingService> sl = ServiceLoader.load(ClassloadingService.class);
-            classloadingService = sl.iterator().next();
+            classloadingService = classloadingServices.iterator().next();
         } catch (Exception ex) {
             classloadingService = new DefaultClassloadingService();
         }
